#521 Refactor Admin page; add Action Plan Export Tool to tool list
